Move logic out of Exrm-specific module; add distillery module.
